+++ resources/lang/sk/general.php
@@ -42,7 +42,6 @@ return array(
     "share_link"         => "Zdieľať URL adresu",
     "copy"               => "kopírovať",
     "copied_to_clipboard"=> "Odkaz bol skopírovaný do schránky",
-    'minute'             => ':count minúta|:count minúty|:count minút',
-    'reading_time'       => 'Čas čítania',
-    "citation_copied"    => "Citácia bola skopírovaná"
+    "citation_copied"    => "Citácia bola skopírovaná",
+    "minute"             => ":count minúta|:count minúty|:count minút",
 );