+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Fixed
+- Fix patternlib page
 
 ## [2.20.0] - 2021-02-26
 ### Added