+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file[^1].
 ### Fixed
 - Fix linked-combos links for item relationships
 - Fix form for articles and collections
+- Fix patternlib page
 
 ## [2.20.0] - 2021-02-26
 ### Added